backend/requirements.txt
# ========================================
# DEPENDANCES PRINCIPALES DU BACKEND
# ========================================

# Framework Web
Flask==3.0.0
Flask-CORS==4.0.0

# Variables d'environnement
python-dotenv==1.0.0


# ========================================
# GOOGLE ADS API
# ========================================

# Google Ads API Client
google-ads==28.4.1
google-api-core>=2.18.0

# ========================================
# GOOGLE SHEETS API
# ========================================

# Google API Client pour Sheets
google-api-python-client==2.108.0

# ========================================
# GOOGLE ANALYTICS DATA API (GA4)
# ========================================

google-analytics-data==0.18.16

# Authentification Google
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# ========================================
# REQUESTS ET UTILITAIRES
# ========================================

# Requêtes HTTP
requests==2.31.0
beautifulsoup4==4.12.2

# ========================================
# DEPENDANCES DES SCRIPTS D'AUTOMATISATION
# ========================================

# Selenium pour l'automatisation web
selenium==4.15.2
webdriver-manager==4.0.1

# ========================================
# UTILITAIRES DE DÉVELOPPEMENT
# ========================================

# Logging et debugging
colorlog==6.8.0

# Validation des données
marshmallow==3.20.1

# ========================================
# RAPPORTS PPTX
# ========================================

# Génération de présentations PowerPoint
python-pptx>=0.6.21

# Graphiques pour les rapports
matplotlib>=3.8.0
plotly>=6.0.0
kaleido>=1.0.0

# Calcul de dates relatives (M-1, M-2)
python-dateutil>=2.8.0

# ========================================
# SÉCURITÉ
# ========================================

# Cryptographie pour les tokens
cryptography==41.0.7

# ========================================
# OPTIMISATION ET PERFORMANCE
# ========================================

# Cache et optimisation
redis==5.0.1

# ========================================
# TESTS (optionnel)
# ========================================

# Tests unitaires
pytest==7.4.3
pytest-flask==1.3.0
psutil==5.9.6

# ========================================
# DÉPLOIEMENT
# ========================================

# Serveur WSGI pour production
gunicorn==21.2.0

# ========================================
# NOTES D'INSTALLATION
# ========================================

# Pour installer toutes les dépendances :
# pip install -r requirements.txt

# Pour installer uniquement les dépendances de production :
# pip install -r requirements.txt --no-dev

# Versions Python supportées : 3.8+ 